Refactor Eigen Portfolio strategy to inherit from PortfolioBase and streamline initialization
